refactor(vamm): Move Vamm types
Move `SwapDirection` and `VammState` to Pallet Types in `lib.rs`
