+++ frame/vamm/src/lib.rs
@@ -159,6 +159,35 @@ pub mod pallet {
 	type VammIdOf<T> = <T as Config>::VammId;
 	type VammStateOf<T> = VammState<BalanceOf<T>, TimestampOf<T>>;
 
+	/// Represents the direction a of a position.
+	#[derive(Encode, Decode, MaxEncodedLen, TypeInfo)]
+	pub enum SwapDirection {
+		Add,
+		Remove,
+	}
+
+	/// Data relating to the state of a virtual market.
+	#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Copy, PartialEq, Debug)]
+	pub struct VammState<Balance, Timestamp> {
+		/// The total amount of base asset present in the virtual market.
+		pub base_asset_reserves: Balance,
+
+		/// The total amount of quote asset present in the virtual market.
+		pub quote_asset_reserves: Balance,
+
+		/// The magnitude of the quote asset reserve.
+		pub peg_multiplier: Balance,
+
+		/// Whether this market is deprecated or not.
+		///
+		/// This variable function as a signal to allow pallets who uses the Vamm to
+		/// set a market as "operating as normal" or "not to be used anymore".  If
+		/// the value is `None` it means the market is operating as normal, but if
+		/// the value is `Some(timestamp)` it means the market is deprecated and the
+		/// deprecation will take (or took) effect at the time `timestamp`.
+		pub deprecated: Option<Timestamp>,
+	}
+
 	// ----------------------------------------------------------------------------------------------------
 	//                                           Runtime  Storage
 	// ----------------------------------------------------------------------------------------------------
